Add files via upload
Check figma link here.. still a work in progress. 

https://www.figma.com/file/Cl9ZOrUzEBEryD5Ppbv8La/Untitled?node-id=0%3A1

Basically the project would be launching my company website (which is called Get Down a upcoming condom company). As i don't think i have the skills to do full eccomerce store, I would like to do a launch page for people to sign up pre launch.. which goes to either a klaviyo or mail chimp list? or if that's to hard even a google sheet. 

Not sure how hard an instagram feed would be or if i just keep that element static.

the wave section.. i wanted to try do some movement, but if too hard just shapes.